+++ CHANGELOG.md
@@ -7,3 +7,12 @@ version links.
 [commits]: https://github.com/seanpdoyle/action_view-attributes
 
 ## main
+
+*   Ensure that attribute values like [data-action](https://stimulus.hotwired.dev/reference/actions)
+    descriptor syntax aren't escaped too many times when transformed to Token Lists
+
+    *Sean Doyle*
+
+## 0.1.0 (Feb 5, 2023)
+
+*   Extract out of [attributes_and_token_lists](https://github.com/seanpdoyle/attributes_and_token_lists)